+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [#911](https://github.com/spegel-org/spegel/pull/911) Enforce use of request contexts and fix response closing.
+- [#914](https://github.com/spegel-org/spegel/pull/914) Fix OCI client header parsing and improve tests.
 
 ### Security